// admin-dashboard/store/analytics/mutation-types.ts
export enum MutationTypes {
  SET_USER = "SET_USER",
  SET_SIGN_IN_DATA = "SET_SIGN_IN_DATA",
  SET_SIGN_UP_DATA = "SET_SIGN_UP_DATA",
  SET_HAS_USER = "SET_HAS_USER",
  SET_USER_COUNT_ANALYTICS = "SET_USER_COUNT_ANALYTICS",
  SET_FILM_COUNT_ANALYTICS = "SET_FILM_COUNT_ANALYTICS",
  SET_ADMIN_COUNT_ANALYTICS = "SET_ADMIN_COUNT_ANALYTICS",
  SET_USER_CUMULATIVE_COUNT_ANALYTICS = "SET_USER_CUMULATIVECOUNT_ANALYTICS",
  SET_FILM_CUMULATIVE_COUNT_ANALYTICS = "SET_FILM_CUMULATIVE_COUNT_ANALYTICS",
  SET_ADMIN_CUMULATIVE_COUNT_ANALYTICS = "SET_ADMIN_CUMULATIVE_COUNT_ANALYTICS",
  SET_USER_DATE_COUNT_ANALYTICS = "SET_USER_DATE_COUNT_ANALYTICS",
  SET_FILM_DATE_COUNT_ANALYTICS = "SET_FILM_DATE_COUNT_ANALYTICS",
  SET_ADMIN_DATE_COUNT_ANALYTICS = "SET_ADMIN_DATE_COUNT_ANALYTICS",
  SET_LOADING = "SET_LOADING",
}